refactor(deepseek_ocr): specialize family implementation
Consolidate deepseek_ocr model construction and weight mapping under its owned package. Remove unused family-local duplicates while preserving the existing plugin API.
