--- python/tensorrt_model_connect/families/deepseek_ocr/standard_decoder_builder.py
@@ -1,13 +1,0 @@
-"""Compatibility shim for the shared default decoder builder."""
-
-from .default_decoder import (
-    _apply_norm,
-    _mark_debug_output,
-    build_standard_decoder_engine,
-)
-
-__all__ = [
-    "_apply_norm",
-    "_mark_debug_output",
-    "build_standard_decoder_engine",
-]
