+++ tools/families/deepseek_ocr/__init__.py
@@ -1,0 +1,1 @@
+"""Family-owned development tools."""
